+++ examples/example_fortranWrapper_saxpy.f90
@@ -6,10 +6,10 @@
 
 program example_fortranWrapper_saxpy
 
-    integer, parameter :: n = 10, incx = 1, incy = 1
+    integer(8), parameter :: n = 10, incx = 1, incy = 1
     real, parameter :: alpha = 2.e0
     real :: x(n), y(n)
-    integer :: i
+    integer(8) :: i
 
     external :: saxpy
